Add files via upload
This was a code i researched and was able to come up with a couple of years back. Business problem was Excel developer and some VB code made this a report that took hours to pull each month especially if Excel crashed. This code reduced it down to 5 minutes a month in an SSRS report.